refactor(tabs): split lnbitsServiceLocal into cohesive modules
Closes #211.

lnbitsServiceLocal.ts was 1297 lines, the only file over the 1000-line
max-lines guardrail, and was grandfathered with a per-file ESLint override.

Split it by responsibility into tabs/src/services/lnbits/: config, auth,
cache, wallets, users, payments and rewards. lnbitsServiceLocal.ts is now a
barrel that re-exports the same public API, so every existing import keeps
working unchanged.

The module-level dedup state moves into cache.ts. pendingUsersRequest and
pendingWalletRequests now live on a single shared pendingRequests object so
the owning modules can reassign the in-flight users promise across module
boundaries, which a re-exported let binding does not allow.

Code is moved verbatim; no behaviour change. Removes the max-lines override
from .eslintrc.json.
