+++ tabs/src/services/lnbits/config.ts
@@ -1,0 +1,6 @@
+// LNBits API is documented here:
+// https://demo.lnbits.com/docs/
+
+export const userName = process.env.REACT_APP_LNBITS_USERNAME;
+export const password = process.env.REACT_APP_LNBITS_PASSWORD;
+export const nodeUrl = process.env.REACT_APP_LNBITS_NODE_URL;
